+++ Emulator/Peripherals/ControlPort.cpp
@@ -41,13 +41,13 @@ ControlPort::_inspect()
 void
 ControlPort::_dump()
 {
-    msg("         device: %d (%s)\n",
-        device,
+    msg("         device: %ld (%s)\n",
+        (long)device,
         device == CPD_NONE ? "CPD_NONE" :
         device == CPD_MOUSE ? "CPD_MOUSE" :
         device == CPD_JOYSTICK ? "CPD_JOYSTICK" : "???");
-    msg("  mouseCounterX: %d\n", mouseCounterX);
-    msg("  mouseCounterY: %d\n", mouseCounterY);
+    msg("  mouseCounterX: %lld\n", mouseCounterX);
+    msg("  mouseCounterY: %lld\n", mouseCounterY);
 }
 
 u16
